backend/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.3
django-cors-headers==3.13.0
django-debug-toolbar==3.7.0
django-redis==5.2.0
djangorestframework==3.14.0
kombu==5.2.4
packaging==21.3
prompt-toolkit==3.0.33
psycopg2==2.9.5
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.6
redis==4.3.5
six==1.16.0
sqlparse==0.4.3
vine==5.0.0
wcwidth==0.2.5
